Revert 78 feature/member

+++ src/main/java/com/nhnacademy/byeol23backend/couponset/couponpolicy/domain/dto/CouponPolicyCreateRequest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record CouponPolicyCreateRequest(
         String policyName,
-        BigDecimal criterionPrice,
+        Long criterionPrice,
         Integer discountRate,
         BigDecimal discountLimit,
         BigDecimal discountAmount
